needed to call updateFavoriteButtons(); everyewhere
